refactor: 컴포넌트를 _common과 _only로 구분
