+++ app/layout.tsx
@@ -4,9 +4,9 @@ import type { Metadata } from "next";
 import { Black_Han_Sans, Noto_Sans_KR } from "next/font/google";
 import localFont from "next/font/local";
 
-import { Header } from "./_common/components/Header";
-import { Providers } from "./_common/Providers";
 import { classNames } from "./_common/utils/classNames";
+import { Header } from "./_only/components/Header";
+import { Providers } from "./_only/Providers";
 
 const notoSansKr = Noto_Sans_KR({
   subsets: ["latin"],
